panels: extract some metric rows to a subcomponents
