+++ src/view/panels/animation/AnimationCancelHistory.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
   import {
-    CafFacts,
     type TCancelAnimationFrameHistory,
   } from '../../../wrapper/AnimationWrapper.ts';
   import {
@@ -10,11 +9,8 @@
   import { compareByFieldOrder } from '../shared/comparator.ts';
   import Variable from '../../shared/Variable.svelte';
   import ColumnSortable from '../shared/ColumnSortable.svelte';
-  import CellBreakpoint from '../shared/CellBreakpoint.svelte';
-  import CellBypass from '../shared/CellBypass.svelte';
-  import CellFacts from '../shared/CellFacts.svelte';
-  import CellCallstack from '../shared/CellCallstack.svelte';
   import { useConfigState } from '../../../state/config.state.svelte.ts';
+  import AnimationCancelHistoryMetric from './AnimationCancelHistoryMetric.svelte';
 
   let {
     cafHistory,
@@ -79,21 +75,7 @@
 
   <tbody>
     {#each sortedMetrics as metric (metric.traceId)}
-      <tr class="t-zebra">
-        <td class="wb-all">
-          <CellCallstack
-            trace={metric.trace}
-            traceDomain={metric.traceDomain}
-          />
-        </td>
-        <td class="ta-c">
-          <CellFacts facts={metric.facts} factsMap={CafFacts} />
-        </td>
-        <td class="ta-c"><Variable value={metric.calls} /></td>
-        <td class="ta-c">{metric.handler}</td>
-        <td><CellBypass traceId={metric.traceId} /></td>
-        <td><CellBreakpoint traceId={metric.traceId} /></td>
-      </tr>
+      <AnimationCancelHistoryMetric {metric} />
     {/each}
   </tbody>
 </table>
