+++ src/view/panels/idle/IdleCallbackCancelHistory.svelte
@@ -7,12 +7,8 @@
   import { compareByFieldOrder } from '../shared/comparator.ts';
   import Variable from '../../shared/Variable.svelte';
   import ColumnSortable from '../shared/ColumnSortable.svelte';
-  import CellBreakpoint from '../shared/CellBreakpoint.svelte';
-  import CellBypass from '../shared/CellBypass.svelte';
-  import { CicFacts } from '../../../wrapper/IdleWrapper.ts';
-  import CellFacts from '../shared/CellFacts.svelte';
-  import CellCallstack from '../shared/CellCallstack.svelte';
   import { useConfigState } from '../../../state/config.state.svelte.ts';
+  import IdleCallbackCancelHistoryMetric from './IdleCallbackCancelHistoryMetric.svelte';
 
   let {
     cicHistory,
@@ -76,21 +72,7 @@
 
   <tbody>
     {#each sortedMetrics as metric (metric.traceId)}
-      <tr class="t-zebra">
-        <td class="wb-all">
-          <CellCallstack
-            trace={metric.trace}
-            traceDomain={metric.traceDomain}
-          />
-        </td>
-        <td class="ta-c">
-          <CellFacts facts={metric.facts} factsMap={CicFacts} />
-        </td>
-        <td class="ta-c"><Variable value={metric.calls} /></td>
-        <td class="ta-c">{metric.handler}</td>
-        <td><CellBypass traceId={metric.traceId} /></td>
-        <td><CellBreakpoint traceId={metric.traceId} /></td>
-      </tr>
+      <IdleCallbackCancelHistoryMetric {metric} />
     {/each}
   </tbody>
 </table>
