--- src/view/panels/online/Online.svelte
@@ -1,29 +1,0 @@
-<script lang="ts">
-  import {
-    ETimerType,
-    type TOnlineTimerMetrics,
-  } from '../../../wrapper/TimerWrapper.ts';
-  import { compareByDelayThenHandlerDescending } from '../shared/comparator.ts';
-  import OnlineTimers from './OnlineTimers.svelte';
-
-  let { onlineTimers }: { onlineTimers: TOnlineTimerMetrics[] | null } =
-    $props();
-  let timeouts = $derived.by(() =>
-    (onlineTimers || []).filter((o) => o.type === ETimerType.TIMEOUT).sort(
-      compareByDelayThenHandlerDescending,
-    )
-  );
-  let intervals = $derived.by(() =>
-    (onlineTimers || []).filter((o) => o.type === ETimerType.INTERVAL).sort(
-      compareByDelayThenHandlerDescending,
-    )
-  );
-</script>
-
-{#if intervals.length}
-  <OnlineTimers caption="Active Intervals" metrics={intervals} />
-{/if}
-
-{#if timeouts.length}
-  <OnlineTimers caption="Active Timeouts" metrics={timeouts} />
-{/if}
